LayoutTests/accessibility/ios-simulator/form-control-validation-message-expected.txt

This tests that we are exposing form validation message through hint on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.helpText is "AXHint: "
PASS input.helpText is "AXHint: Fill out this field"
PASS successfullyParsed is true

TEST COMPLETE
Required text input:
